+++ GameEngine/Random.cpp
@@ -2,7 +2,7 @@
 
 namespace GameEngine {
 	Random::Random() : screenHeight(0), screenWidth(0) {
-		_mt.seed(std::chrono::system_clock::now().time_since_epoch().count());
+		_mt.seed((unsigned long)std::chrono::system_clock::now().time_since_epoch().count());
 	}
 
 
@@ -10,7 +10,7 @@ namespace GameEngine {
 	}
 
 	float Random::random(float lowerBound, float upperBound) {
-		std::uniform_real_distribution<double> dist(lowerBound, upperBound);
+		std::uniform_real_distribution<float> dist(lowerBound, upperBound);
 		return dist(_mt);
 	}
 
@@ -25,7 +25,7 @@ namespace GameEngine {
 
 	int Random::randomIntDist(std::discrete_distribution<int> dist) {
 		//std::discrete_distribution<int> distb{0,1,1,1};
-		std::mt19937 gen(std::chrono::system_clock::now().time_since_epoch().count());
+		std::mt19937 gen((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
 		return dist(gen);
 	}
 }
